Control.Concurrent.JobPool

Description

This module allows the management of a multiple Async jobs which are grouped by an 'Ord group => group' type.

Documentation

data JobPool group m a

JobPool allows to submit asynchronous jobs, wait for their completion or cancel. Jobs are grouped, each group can be cancelled separately.

data Job group m a

An asynchronous job which belongs to some group and its exception handler.

Constructors

Job

Fields

(m a)

job

(SomeException -> m a)

error handler

group

job group

String

thread label

withJobPool :: forall group m a b. (MonadAsync m, MonadThrow m, MonadLabelledSTM m) => (JobPool group m a -> m b) -> m b

forkJob :: forall group m a. (MonadAsync m, MonadMask m, Ord group) => JobPool group m a -> Job group m a -> m ()

readSize :: MonadSTM m => JobPool group m a -> STM m Int

readGroupSize :: (MonadSTM m, Eq group) => JobPool group m a -> group -> STM m Int

waitForJob :: MonadSTM m => JobPool group m a -> STM m a

Wait for next successfully completed job. Unlike wait it will not throw if a job errors.

cancelGroup :: (MonadAsync m, Eq group) => JobPool group m a -> group -> m ()

Cancel all threads in a given group. Blocks until all threads terminated.