Network.Mux.Bearer.AttenuatedChannel

Documentation

data AttenuatedChannel m

An AttenuatedChannel supports:

attenuation applied after reading a message from QueueChannel;

two-way close handshake with 120s timeout. Read side is closed as soon as an internal MsgClose is received, write side has to be closed explicitly.

Constructors

AttenuatedChannel

Fields

acRead :: m ByteString

acWrite :: ByteString -> m ()

acClose :: m ()

type Size = Int64

data SuccessOrFailure

Constructors

Success

Failure

data Attenuation

Attenuation of a channel.

Constructors

Attenuation

Fields

aReadAttenuation :: Time -> Size -> (DiffTime, SuccessOrFailure)

aWriteAttenuation :: Maybe Int

newConnectedAttenuatedChannelPair :: forall m. (MonadDelay m, MonadLabelledSTM m, MonadTimer m, MonadThrow m, MonadThrow (STM m)) => Tracer m AttenuatedChannelTrace -> Tracer m AttenuatedChannelTrace -> Attenuation -> Attenuation -> STM m (AttenuatedChannel m, AttenuatedChannel m)

Create a pair of connected AttenuatedChannels.

attenuationChannelAsMuxBearer :: forall m. (MonadThrow m, MonadMonotonicTime m) => SDUSize -> DiffTime -> Tracer m MuxTrace -> AttenuatedChannel m -> MuxBearer m

Trace

data AttenuatedChannelTrace

Constructors

AttChannLocalClose Bool

AttChannRemoteClose

Utils

resourceVanishedIOError :: String -> String -> IOError