Network.Mux.Bearer.Socket

Documentation

socketAsMuxBearer ∷ DiffTime → Tracer IO MuxTrace → Socket → MuxBearer IO

Create MuxBearer from a socket.

On Windows Async operations are used to read and write from a socket. This means that the socket must be associated with the I/O completion port with associateWithIOCompletionPort.

Note: IOExceptions thrown by sendAll and recv are wrapped in MuxError.