Network.Mux.Compat

Documentation

muxStart :: forall m mode a b. (MonadAsync m, MonadFork m, MonadLabelledSTM m, Alternative (STM m), MonadThrow (STM m), MonadTimer m, MonadMask m) => Tracer m MuxTrace -> MuxApplication mode m a b -> MuxBearer m -> m ()

Mux bearers

data MuxBearer m

Low level access to underlying socket or pipe. There are three smart constructors:

socketAsMuxBearer

pipeAsMuxBearer

Test.Mux.queuesAsMuxBearer

newtype MakeBearer m fd

Constructors

MakeBearer

Fields

getBearer :: DiffTime -> Tracer m MuxTrace -> fd -> m (MuxBearer m)

Defining MuxApplications

data MuxMode where

Constructors

InitiatorMode :: MuxMode

ResponderMode :: MuxMode

InitiatorResponderMode :: MuxMode

type family HasInitiator (mode :: MuxMode) :: Bool where ...

Equations

HasInitiator InitiatorMode = True

HasInitiator ResponderMode = False

HasInitiator InitiatorResponderMode = True

type family HasResponder (mode :: MuxMode) :: Bool where ...

Equations

HasResponder InitiatorMode = False

HasResponder ResponderMode = True

HasResponder InitiatorResponderMode = True

newtype MuxApplication (mode :: MuxMode) m a b

Constructors

MuxApplication [MuxMiniProtocol mode m a b]

data MuxMiniProtocol (mode :: MuxMode) m a b

Constructors

MuxMiniProtocol

Fields

miniProtocolNum :: !MiniProtocolNum

miniProtocolLimits :: !MiniProtocolLimits

miniProtocolRun :: !(RunMiniProtocol mode m a b)

data RunMiniProtocol (mode :: MuxMode) m a b where

Constructors

InitiatorProtocolOnly :: (Channel m -> m (a, Maybe ByteString)) -> RunMiniProtocol InitiatorMode m a Void

ResponderProtocolOnly :: (Channel m -> m (b, Maybe ByteString)) -> RunMiniProtocol ResponderMode m Void b

InitiatorAndResponderProtocol :: (Channel m -> m (a, Maybe ByteString)) -> (Channel m -> m (b, Maybe ByteString)) -> RunMiniProtocol InitiatorResponderMode m a b

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

newtype MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue :: Int

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

data MiniProtocolDir

Constructors

InitiatorDir

ResponderDir

Errors

data MuxError

Error type used in across the mux layer.

Constructors

MuxError

Fields

errorType :: !MuxErrorType

errorMsg :: !String

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

MuxSDUWriteTimeout
thrown when writing a single SDU takes too long

MuxShutdown !(Maybe MuxErrorType)
Result of runMiniProtocol's completionAction in case of an error or mux being closed while a mini-protocol was still running, this is not a clean exit.

MuxCleanShutdown
Mux stopped by stopMux

Tracing

traceMuxBearerState :: Tracer m MuxTrace -> MuxBearerState -> m ()

data MuxBearerState

Constructors

Mature
MuxBearer has successfully completed the handshake.

Dead
MuxBearer is dead and the underlying bearer has been closed.

data MuxTrace

Enumeration of Mux events that can be traced.

Constructors

MuxTraceRecvHeaderStart

MuxTraceRecvHeaderEnd MuxSDUHeader

MuxTraceRecvDeltaQObservation MuxSDUHeader Time

MuxTraceRecvDeltaQSample Double Int Int Double Double Double Double String

MuxTraceRecvStart Int

MuxTraceRecvEnd Int

MuxTraceSendStart MuxSDUHeader

MuxTraceSendEnd

MuxTraceState MuxBearerState

MuxTraceCleanExit MiniProtocolNum MiniProtocolDir

MuxTraceExceptionExit MiniProtocolNum MiniProtocolDir SomeException

MuxTraceChannelRecvStart MiniProtocolNum

MuxTraceChannelRecvEnd MiniProtocolNum Int

MuxTraceChannelSendStart MiniProtocolNum Int

MuxTraceChannelSendEnd MiniProtocolNum

MuxTraceHandshakeStart

MuxTraceHandshakeClientEnd DiffTime

MuxTraceHandshakeServerEnd

forall e.Exception e => MuxTraceHandshakeClientError e DiffTime

forall e.Exception e => MuxTraceHandshakeServerError e

MuxTraceSDUReadTimeoutException

MuxTraceSDUWriteTimeoutException

MuxTraceStartEagerly MiniProtocolNum MiniProtocolDir

MuxTraceStartOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceStartedOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceTerminating MiniProtocolNum MiniProtocolDir

MuxTraceStopping

MuxTraceStopped

MuxTraceTCPInfo StructTCPInfo Word16

data WithMuxBearer peerid a

Type used for tracing mux events.

Constructors

WithMuxBearer

Fields

wmbPeerId :: !peerid

A tag that should identify a specific mux bearer.

wmbEvent :: !a