Network.Mux.Trace

Documentation

data Error

Enumeration of error conditions.

Constructors

UnknownMiniProtocol MiniProtocolNum
returned by decodeSDUHeader, thrown by Bearer.

BearerClosed String
thrown by Bearer when received a null byte.

IngressQueueOverRun MiniProtocolNum MiniProtocolDir
thrown by demux when violating maximumIngressQueue byte limit.

InitiatorOnly MiniProtocolNum
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

IOException IOException String
Error thrown by

SDUDecodeError String
return by decodeSDUHeader, thrown by Bearer.

SDUReadTimeout
thrown when reading of a single SDU takes too long

SDUWriteTimeout
thrown when writing a single SDU takes too long

Shutdown (Maybe SomeException) Status
Result of runMiniProtocol's completionAction in case of an error or mux being closed while a mini-protocol was still running, this is not a clean exit.

handleIOException :: MonadThrow m => String -> IOException -> m a

Handler for Errors which wraps them in Error.

It is used various Bearer implementations: * socketAsBearer * pipeAsBearer

data Trace

Enumeration of Mux events that can be traced.

Constructors

TraceRecvHeaderStart

TraceRecvHeaderEnd SDUHeader

TraceRecvDeltaQObservation SDUHeader Time

TraceRecvDeltaQSample Double Int Int Double Double Double Double String

TraceRecvStart Int

TraceRecvEnd Int

TraceSendStart SDUHeader

TraceSendEnd

TraceState BearerState

TraceCleanExit MiniProtocolNum MiniProtocolDir

TraceExceptionExit MiniProtocolNum MiniProtocolDir SomeException

TraceChannelRecvStart MiniProtocolNum

TraceChannelRecvEnd MiniProtocolNum Int

TraceChannelSendStart MiniProtocolNum Int

TraceChannelSendEnd MiniProtocolNum

TraceHandshakeStart

TraceHandshakeClientEnd DiffTime

TraceHandshakeServerEnd

forall e.Exception e => TraceHandshakeClientError e DiffTime

forall e.Exception e => TraceHandshakeServerError e

TraceSDUReadTimeoutException

TraceSDUWriteTimeoutException

TraceStartEagerly MiniProtocolNum MiniProtocolDir

TraceStartOnDemand MiniProtocolNum MiniProtocolDir

TraceStartOnDemandAny MiniProtocolNum MiniProtocolDir

TraceStartedOnDemand MiniProtocolNum MiniProtocolDir

TraceTerminating MiniProtocolNum MiniProtocolDir

TraceStopping

TraceStopped

TraceTCPInfo StructTCPInfo Word16

data BearerState

Constructors

Mature
Bearer has successfully completed the handshake.

Dead
Bearer is dead and the underlying bearer has been closed.

data WithBearer peerid a

Type used for tracing mux events.

Constructors

WithBearer

Fields

wbPeerId :: !peerid

A tag that should identify a specific mux bearer.

wbEvent :: !a

data TraceLabelPeer peerid a

A peer label for use in Tracers. This annotates tracer output as being associated with a given peer identifier.

Constructors

TraceLabelPeer peerid a