Network.Mux

Defining Mux protocol bundles

newMux ∷ MonadSTM m ⇒ MiniProtocolBundle mode → m (Mux mode m)

data Mux (mode ∷ MuxMode) m

data MuxMode where

Constructors

InitiatorMode ∷ MuxMode

ResponderMode ∷ MuxMode

InitiatorResponderMode ∷ MuxMode

type family HasInitiator (mode ∷ MuxMode) ∷ Bool where ...

Equations

HasInitiator InitiatorMode = True

HasInitiator ResponderMode = False

HasInitiator InitiatorResponderMode = True

type family HasResponder (mode ∷ MuxMode) ∷ Bool where ...

Equations

HasResponder InitiatorMode = False

HasResponder ResponderMode = True

HasResponder InitiatorResponderMode = True

newtype MiniProtocolBundle (mode ∷ MuxMode)

Application run by mux layer.

enumeration of client application, e.g. a wallet application communicating with a node using ChainSync and TxSubmission protocols; this only requires to run client side of each protocol.

enumeration of server applications: this application type is mostly useful tests.

enumeration of both client and server applications, e.g. a full node serving downstream peers using server side of each protocol and getting updates from upstream peers using client side of each of the protocols.

Constructors

MiniProtocolBundle [MiniProtocolInfo mode]

data MiniProtocolInfo (mode ∷ MuxMode)

Constructors

MiniProtocolInfo

Fields

miniProtocolNum ∷ !MiniProtocolNum

miniProtocolDir ∷ !(MiniProtocolDirection mode)

miniProtocolLimits ∷ !MiniProtocolLimits

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolDirection (mode ∷ MuxMode) where

Constructors

InitiatorDirectionOnly ∷ MiniProtocolDirection InitiatorMode

ResponderDirectionOnly ∷ MiniProtocolDirection ResponderMode

InitiatorDirection ∷ MiniProtocolDirection InitiatorResponderMode

ResponderDirection ∷ MiniProtocolDirection InitiatorResponderMode

data MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue ∷ !Int

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

Running the Mux

runMux ∷ ∀ m mode. (MonadAsync m, MonadCatch m, MonadFork m, MonadLabelledSTM m, MonadThrow (STM m), MonadTime m, MonadTimer m, MonadMask m) ⇒ Tracer m MuxTrace → Mux mode m → MuxBearer m → m ()

runMux starts a mux bearer for the specified protocols corresponding to one of the provided Versions.

Isometric flow control: analysis of head-of-line blocking of the ingress side of the multiplexer

For each mini-protocol (enumeratated by ptcl), mux will create two channels. One for initiator and one for the responder. Each channel will use a single Wanton. When it is filled, it is put in a common queue tsrQueue. This means that the queue is bound by 2 * |ptcl|. Every side of a mini-protocol is served by a single Wanton: when an applicaiton sends data, the channel will try to put it into the Wanton (which might block). Wantons are taken from the tsrQueue queue by one of mux threads. This elimnates head of line blocking: each mini-protocol thread can block on puting more bytes into its Wanton, but it cannot block the other mini-protocols or the thread that is reading the tsrQueue queue. This is ensured since the muxChannel will put only a non-empty Wanton to the tsrQueue queue, and on such wantons the queue is never blocked. This means that the only way the queue can block is when its empty, which means that none of the mini-protocols wanted to send. The egress part will read a Wanton, take a fixed amount of bytes encode them in as an MuxSDU; if there are leftovers it will put them back in the Wanton and place it at the end of the queue (reading and writting to it will happen in a single STM transaction which assures that the order of requests from a mini-protocol is preserved.

Properties:

at any given time the tsrQueue contains at most one TranslocationServiceRequest from a given mini-protocol of the given MiniProtocolDir, thus the queue contains at most 2 * |ptcl| translocation requests.

at any given time each TranslocationServiceRequest contains a non-empty Wanton

data MuxBearer m

Low level access to underlying socket or pipe. There are three smart constructors:

socketAsMuxBearer

pipeAsMuxBearer

Test.Mux.queuesAsMuxBearer

runMiniProtocol ∷ ∀ mode m a. (MonadSTM m, MonadThrow m, MonadThrow (STM m)) ⇒ Mux mode m → MiniProtocolNum → MiniProtocolDirection mode → StartOnDemandOrEagerly → (Channel m → m (a, Maybe ByteString)) → m (STM m (Either SomeException a))

Arrange to run a protocol thread (for a particular MiniProtocolNum and MiniProtocolDirection) to interact on this protocol's Channel.

The protocol thread can either be started eagerly or on-demand:

With StartEagerly, the thread is started promptly. This is appropriate for mini-protocols where the opening message may be sent by this thread.

With StartOnDemand, the thread is not started until the first data is received for this mini-protocol. This is appropriate for mini-protocols where the opening message is sent by the remote peer.

The result is a STM action to block and wait on the protocol completion. It is safe to call this completion action multiple times: it will always return the same result once the protocol thread completes. Incase the Mux has stopped, either due to an exception or because of a call to muxStop a `Left MuxError` will be returned from the STM action.

It is an error to start a new protocol thread while one is still running, for the same MiniProtocolNum and MiniProtocolDirection. This can easily be avoided by using the STM completion action to wait for the previous one to finish.

It is safe to ask to start a protocol thread before runMux. In this case the protocol thread will not actually start until runMux is called, irrespective of the StartOnDemandOrEagerly value.

data StartOnDemandOrEagerly

Constructors

StartOnDemand

StartEagerly

stopMux ∷ MonadSTM m ⇒ Mux mode m → m ()

Shut down the mux. This will cause runMux to return. It does not wait for any protocol threads to finish, so you should do that first if necessary.

Monitoring

miniProtocolStateMap ∷ MonadSTM m ⇒ Mux mode m → Map (MiniProtocolNum, MiniProtocolDir) (STM m MiniProtocolStatus)

muxStopped ∷ MonadSTM m ⇒ Mux mode m → STM m (Maybe SomeException)

Await until mux stopped.

Errors

data MuxError

Error type used in accross the mux layer.

Constructors

MuxError

Fields

errorType ∷ !MuxErrorType

errorMsg ∷ !String

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

MuxSDUWriteTimeout
thrown when writing a single SDU takes too long

MuxShutdown !(Maybe MuxErrorType)
Result of runMiniProtocol's completionAction in case of an error or mux being closed while a mini-protocol was still running, this is not a clean exit.

MuxCleanShutdown
Mux stopped by stopMux

Tracing

traceMuxBearerState ∷ Tracer m MuxTrace → MuxBearerState → m ()

data MuxBearerState

Constructors

Mature
MuxBearer has successufully completed the handshake.

Dead
MuxBearer is dead and the underlying bearer has been closed.

data MuxTrace

Enumeration of Mux events that can be traced.

Constructors

MuxTraceRecvHeaderStart

MuxTraceRecvHeaderEnd MuxSDUHeader

MuxTraceRecvDeltaQObservation MuxSDUHeader Time

MuxTraceRecvDeltaQSample Double Int Int Double Double Double Double String

MuxTraceRecvStart Int

MuxTraceRecvEnd Int

MuxTraceSendStart MuxSDUHeader

MuxTraceSendEnd

MuxTraceState MuxBearerState

MuxTraceCleanExit MiniProtocolNum MiniProtocolDir

MuxTraceExceptionExit MiniProtocolNum MiniProtocolDir SomeException

MuxTraceChannelRecvStart MiniProtocolNum

MuxTraceChannelRecvEnd MiniProtocolNum Int

MuxTraceChannelSendStart MiniProtocolNum Int

MuxTraceChannelSendEnd MiniProtocolNum

MuxTraceHandshakeStart

MuxTraceHandshakeClientEnd DiffTime

MuxTraceHandshakeServerEnd

∀ e.Exception e ⇒ MuxTraceHandshakeClientError e DiffTime

∀ e.Exception e ⇒ MuxTraceHandshakeServerError e

MuxTraceSDUReadTimeoutException

MuxTraceSDUWriteTimeoutException

MuxTraceStartEagerly MiniProtocolNum MiniProtocolDir

MuxTraceStartOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceStartedOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceTerminating MiniProtocolNum MiniProtocolDir

MuxTraceShutdown

data WithMuxBearer peerid a

Type used for tracing mux events.

Constructors

WithMuxBearer

Fields

wmbPeerId ∷ !peerid

A tag that should identify a specific mux bearer.

wmbEvent ∷ !a