Network.Mux

Documentation

muxStart ∷ ∀ m appType a b. (MonadAsync m, MonadCatch m, MonadFork m, MonadSTM m, MonadThrow m, MonadThrow (STM m), MonadTime m, MonadTimer m, MonadMask m, Eq (Async m ())) ⇒ Tracer m MuxTrace → MuxApplication appType m a b → MuxBearer m → m ()

muxStart starts a mux bearer for the specified protocols corresponding to one of the provided Versions.

Isometric flow control: analysis of head-of-line blocking of the ingress side of the multiplexer

For each mini-protocol (enumeratated by ptcl), mux will create two channels. One for initiator and one for the responder. Each channel will use a single Wanton. When it is filled, it is put in a common queue tsrQueue. This means that the queue is bound by 2 * |ptcl|. Every side of a mini-protocol is served by a single Wanton: when an applicaiton sends data, the channel will try to put it into the Wanton (which might block). Wantons are taken from the tsrQueue queue by one of mux threads. This elimnates head of line blocking: each mini-protocol thread can block on puting more bytes into its Wanton, but it cannot block the other mini-protocols or the thread that is reading the tsrQueue queue. This is ensured since the muxChannel will put only a non-empty Wanton to the tsrQueue queue, and on such wantons the queue is never blocked. This means that the only way the queue can block is when its empty, which means that none of the mini-protocols wanted to send. The egress part will read a Wanton, take a fixed amount of bytes encode them in as an MuxSDU; if there are leftovers it will put them back in the Wanton and place it at the end of the queue (reading and writting to it will happen in a single STM transaction which assures that the order of requests from a mini-protocol is preserved.

Properties:

at any given time the tsrQueue contains at most one TranslocationServiceRequest from a given mini-protocol of the given MiniProtocolMode, thus the queue contains at most 2 * |ptcl| translocation requests.

at any given time each TranslocationServiceRequest contains a non-empty Wanton

Mux bearers

data MuxBearer m

Low level access to underlying socket or pipe. There are three smart constructors:

socketAsMuxBearer

pipeAsMuxBearer

Test.Mux.queuesAsMuxBearer

Defining MuxApplications

data AppType where

Constructors

InitiatorApp ∷ AppType

ResponderApp ∷ AppType

InitiatorAndResponderApp ∷ AppType

type family HasInitiator (appType ∷ AppType) ∷ Bool where ...

Equations

HasInitiator InitiatorApp = True

HasInitiator ResponderApp = False

HasInitiator InitiatorAndResponderApp = True

type family HasResponder (appType ∷ AppType) ∷ Bool where ...

Equations

HasResponder InitiatorApp = False

HasResponder ResponderApp = True

HasResponder InitiatorAndResponderApp = True

newtype MuxApplication (appType ∷ AppType) m a b

Application run by mux layer.

enumeration of client application, e.g. a wallet application communicating with a node using ChainSync and TxSubmission protocols; this only requires to run client side of each protocol.

enumeration of server applications: this application type is mostly useful tests.

enumeration of both client and server applications, e.g. a full node serving downstream peers using server side of each protocol and getting updates from upstream peers using client side of each of the protocols.

Constructors

MuxApplication [MuxMiniProtocol appType m a b]

data MuxMiniProtocol (appType ∷ AppType) m a b

Constructors

MuxMiniProtocol

Fields

miniProtocolNum ∷ !MiniProtocolNum

miniProtocolLimits ∷ !MiniProtocolLimits

miniProtocolRun ∷ !(RunMiniProtocol appType m a b)

data RunMiniProtocol (appType ∷ AppType) m a b where

Constructors

InitiatorProtocolOnly ∷ (Channel m → m a) → RunMiniProtocol InitiatorApp m a Void

ResponderProtocolOnly ∷ (Channel m → m b) → RunMiniProtocol ResponderApp m Void b

InitiatorAndResponderProtocol ∷ (Channel m → m a) → (Channel m → m b) → RunMiniProtocol InitiatorAndResponderApp m a b

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue ∷ !Int64

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

data MiniProtocolMode

Constructors

ModeInitiator

ModeResponder

Errors

data MuxError

Error type used in accross the mux layer.

Constructors

MuxError

Fields

errorType ∷ !MuxErrorType

errorMsg ∷ !String

errorStack ∷ !CallStack

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

MuxSDUWriteTimeout
thrown when writing a single SDU takes too long

Tracing

traceMuxBearerState ∷ Tracer m MuxTrace → MuxBearerState → m ()

data MuxBearerState

Constructors

Larval
Newly created MuxBearer.

Connected
MuxBearer is connected to a peer.

Mature
MuxBearer has successufully completed the handshake.

Dying
MuxBearer is in the process of beeing torn down, requests may fail.

Dead
MuxBearer is dead and the underlying bearer has been closed.

data MuxTrace

Enumeration of Mux events that can be traced.

Constructors

MuxTraceRecvHeaderStart

MuxTraceRecvHeaderEnd !MuxSDUHeader

MuxTraceRecvDeltaQObservation !MuxSDUHeader Time

MuxTraceRecvDeltaQSample !Double !Int !Int !Double !Double !Double !Double !String

MuxTraceRecvStart !Int

MuxTraceRecvEnd !Int

MuxTraceSendStart !MuxSDUHeader

MuxTraceSendEnd

MuxTraceState !MuxBearerState

MuxTraceCleanExit !MiniProtocolNum !MiniProtocolMode

MuxTraceExceptionExit !MiniProtocolNum !MiniProtocolMode !SomeException

MuxTraceChannelRecvStart !MiniProtocolNum

MuxTraceChannelRecvEnd !MiniProtocolNum !Int

MuxTraceChannelSendStart !MiniProtocolNum !Int

MuxTraceChannelSendEnd !MiniProtocolNum

MuxTraceHandshakeStart

MuxTraceHandshakeClientEnd !DiffTime

MuxTraceHandshakeServerEnd

Exception e ⇒ MuxTraceHandshakeClientError !e !DiffTime

Exception e ⇒ MuxTraceHandshakeServerError !e

MuxTraceSDUReadTimeoutException

MuxTraceSDUWriteTimeoutException

data WithMuxBearer peerid a

Type used for tracing mux events.

Constructors

WithMuxBearer

Fields

wmbPeerId ∷ !peerid

A tag that should identify a specific mux bearer.

wmbEvent ∷ !a