Network.Mux.DeltaQ.TraceStats

Documentation

step

Arguments

∷ RemoteClockModel
Remote clock timestamp

→ Time
Local clock timestamp

→ Int
the number of octets in the observed outcome

→ StatsA
accumulation state

→ (StatsA, Maybe OneWayDeltaQSample)

data OneWayDeltaQSample

One way measurement for interval. Note that the fields are lazy here so that only calcuation necessary to satisfy strictness of use occurs.

Constructors

OneWaySample

Fields

duration ∷ Double

sumPackets ∷ Int

sumTotalSDU ∷ Int

estDeltaQS ∷ Double

estDeltaQVMean ∷ Double

estDeltaQVVar ∷ Double

estR ∷ Double

sizeDist ∷ String

constructSample ∷ StatsA → OneWayDeltaQSample

data StatsA

Statistics accumulator. Strict evaluation used to keep the memory footprint strictly bounded.

initialStatsA ∷ StatsA

Initial StatsA