Network.Mux.Ingress

Documentation

Ingress Path

                 ●
                 │
                 │ ByteStrings
                 │
        ░░░░░░░░░▼░░░░░░░░░
        ░┌───────────────┐░
        ░│ Bearer.read() │░ Mux Bearer implementation (Socket, Pipes, etc.)
        ░└───────────────┘░
        ░░░░░░░░░│░░░░░░░░░
                ░│░         MuxSDUs
        ░░░░░░░░░▼░░░░░░░░░
        ░┌───────────────┐░
        ░│     demux     │░ For a given Mux Bearer there is a single demux
        ░└───────┬───────┘░ thread reading from the underlying bearer.
        ░░░░░░░░░│░░░░░░░░░
                ░│░
       ░░░░░░░░░░▼░░░░░░░░░░
       ░ ╭────┬────┬─────╮ ░ There is a limited queue (in bytes) for each mode
       ░ │    │    │     │ ░ (responder/initiator) per miniprotocol. Overflowing
       ░ ▼    ▼    ▼     ▼ ░ a queue is a protocol violation and a
       ░│  │ │  │ │  │ │  │░ MuxIngressQueueOverRun exception is thrown
       ░│ci│ │  │ │bi│ │br│░ and the bearer torn down.
       ░│ci│ │cr│ │bi│ │br│░
       ░└──┘ └──┘ └──┘ └──┘░ Every ingress queue has a dedicated thread which will
       ░░│░░░░│░░░░│░░░░│░░░ read application encoded data from its queue.
         │    │    │    │
          application data
         │    │    │    │
         ▼    │    │    ▼
┌───────────┐ │    │  ┌───────────┐
│ muxDuplex │ │    │  │ muxDuplex │
│ Initiator │ │    │  │ Responder │
│ ChainSync │ │    │  │ BlockFetch│
└───────────┘ │    │  └───────────┘
              ▼    ▼
   ┌───────────┐  ┌───────────┐
   │ muxDuplex │  │ muxDuplex │
   │ Responder │  │ Initiator │
   │ ChainSync │  │ BlockFetch│
   └───────────┘  └───────────┘

demux ∷ (MonadSTM m, MonadThrow m, MonadThrow (STM m), HasCallStack) ⇒ DemuxState m → m void

demux runs as a single separate thread and reads complete MuxSDUs from the underlying Mux Bearer and forwards it to the matching ingress queue.

data DemuxState m

Each peer's multiplexer has some state that provides both de-multiplexing details (for despatch of incoming mesages to mini protocols) and for dispatching incoming SDUs. This is shared between the muxIngress and the bearerIngress processes.

Constructors

DemuxState

Fields

dispatchTable ∷ MiniProtocolDispatch m

bearer ∷ MuxBearer m

data MiniProtocolDispatch m

Constructors

MiniProtocolDispatch !(Array MiniProtocolNum (Maybe MiniProtocolIx)) !(Array (MiniProtocolIx, MiniProtocolMode) (MiniProtocolDispatchInfo m))

data MiniProtocolDispatchInfo m

Constructors

MiniProtocolDispatchInfo !(StrictTVar m ByteString) !Int64