Network.Mux.Time

DiffTime

data DiffTime

This is a length of time, as measured by a clock. Conversion functions will treat it as seconds. It has a precision of 10^-12 s.

diffTimeToMicroseconds ∷ DiffTime → Integer

microsecondsToDiffTime ∷ Integer → DiffTime

Compact timestamp

timestampMicrosecondsLow32Bits ∷ Time → Word32

This is a slightly pecluliar operation: it returns the number of microseconds since an arbitrary epoch, modulo 2^32. This number of microseconds wraps every ~35 minutes.

The purpose is to give a compact timestamp (compact to send over the wire) for measuring time differences on the order of seconds or less.