Network.Mux.Trace

Documentation

data MuxError

Error type used in across the mux layer.

Constructors

MuxError

Fields

errorType :: !MuxErrorType

errorMsg :: !String

data MuxErrorType

Enumeration of error conditions.

Constructors

MuxUnknownMiniProtocol
returned by decodeMuxSDUHeader, thrown by MuxBearer.

MuxDecodeError
return by decodeMuxSDUHeader, thrown by MuxBearer.

MuxBearerClosed
thrown by MuxBearer when received a null byte.

MuxIngressQueueOverRun
thrown by demux when violating maximumIngressQueue byte limit.

MuxInitiatorOnly
thrown when data arrives on a responder channel when the mux was set up as an InitiatorApp.

MuxIOException IOException
IOException thrown by

MuxSDUReadTimeout
thrown when reading of a single SDU takes too long

MuxSDUWriteTimeout
thrown when writing a single SDU takes too long

MuxShutdown !(Maybe MuxErrorType)
Result of runMiniProtocol's completionAction in case of an error or mux being closed while a mini-protocol was still running, this is not a clean exit.

MuxCleanShutdown
Mux stopped by stopMux

handleIOException :: MonadThrow m => String -> IOException -> m a

Handler for IOExceptions which wraps them in MuxError.

It is used various MuxBearer implementations: * socketAsMuxBearer * pipeAsMuxBearer

data MuxTrace

Enumeration of Mux events that can be traced.

Constructors

MuxTraceRecvHeaderStart

MuxTraceRecvHeaderEnd MuxSDUHeader

MuxTraceRecvDeltaQObservation MuxSDUHeader Time

MuxTraceRecvDeltaQSample Double Int Int Double Double Double Double String

MuxTraceRecvStart Int

MuxTraceRecvEnd Int

MuxTraceSendStart MuxSDUHeader

MuxTraceSendEnd

MuxTraceState MuxBearerState

MuxTraceCleanExit MiniProtocolNum MiniProtocolDir

MuxTraceExceptionExit MiniProtocolNum MiniProtocolDir SomeException

MuxTraceChannelRecvStart MiniProtocolNum

MuxTraceChannelRecvEnd MiniProtocolNum Int

MuxTraceChannelSendStart MiniProtocolNum Int

MuxTraceChannelSendEnd MiniProtocolNum

MuxTraceHandshakeStart

MuxTraceHandshakeClientEnd DiffTime

MuxTraceHandshakeServerEnd

forall e.Exception e => MuxTraceHandshakeClientError e DiffTime

forall e.Exception e => MuxTraceHandshakeServerError e

MuxTraceSDUReadTimeoutException

MuxTraceSDUWriteTimeoutException

MuxTraceStartEagerly MiniProtocolNum MiniProtocolDir

MuxTraceStartOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceStartedOnDemand MiniProtocolNum MiniProtocolDir

MuxTraceTerminating MiniProtocolNum MiniProtocolDir

MuxTraceStopping

MuxTraceStopped

MuxTraceTCPInfo StructTCPInfo Word16

data MuxBearerState

Constructors

Mature
MuxBearer has successfully completed the handshake.

Dead
MuxBearer is dead and the underlying bearer has been closed.

data WithMuxBearer peerid a

Type used for tracing mux events.

Constructors

WithMuxBearer

Fields

wmbPeerId :: !peerid

A tag that should identify a specific mux bearer.

wmbEvent :: !a

data TraceLabelPeer peerid a

A peer label for use in Tracers. This annotates tracer output as being associated with a given peer identifier.

Constructors

TraceLabelPeer peerid a