Network.Mux.Types

Documentation

newtype MiniProtocolBundle (mode :: MuxMode)

Application run by mux layer.

enumeration of client application, e.g. a wallet application communicating with a node using ChainSync and TxSubmission protocols; this only requires to run client side of each protocol.

enumeration of server applications: this application type is mostly useful tests.

enumeration of both client and server applications, e.g. a full node serving downstream peers using server side of each protocol and getting updates from upstream peers using client side of each of the protocols.

Constructors

MiniProtocolBundle [MiniProtocolInfo mode]

data MiniProtocolInfo (mode :: MuxMode)

Constructors

MiniProtocolInfo

Fields

miniProtocolNum :: !MiniProtocolNum

miniProtocolDir :: !(MiniProtocolDirection mode)

miniProtocolLimits :: !MiniProtocolLimits

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolDirection (mode :: MuxMode) where

Constructors

InitiatorDirectionOnly :: MiniProtocolDirection InitiatorMode

ResponderDirectionOnly :: MiniProtocolDirection ResponderMode

InitiatorDirection :: MiniProtocolDirection InitiatorResponderMode

ResponderDirection :: MiniProtocolDirection InitiatorResponderMode

newtype MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue :: Int

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

data MuxMode where

Constructors

InitiatorMode :: MuxMode

ResponderMode :: MuxMode

InitiatorResponderMode :: MuxMode

type family HasInitiator (mode :: MuxMode) :: Bool where ...

Equations

HasInitiator InitiatorMode = True

HasInitiator ResponderMode = False

HasInitiator InitiatorResponderMode = True

type family HasResponder (mode :: MuxMode) :: Bool where ...

Equations

HasResponder InitiatorMode = False

HasResponder ResponderMode = True

HasResponder InitiatorResponderMode = True

type IngressQueue m = StrictTVar m ByteString

data MiniProtocolIx

The index of a protocol in a MuxApplication, used for array indices

data MiniProtocolDir

Constructors

InitiatorDir

ResponderDir

protocolDirEnum :: MiniProtocolDirection mode -> MiniProtocolDir

data MiniProtocolState mode m

Constructors

MiniProtocolState

Fields

miniProtocolInfo :: MiniProtocolInfo mode

miniProtocolIngressQueue :: IngressQueue m

miniProtocolStatusVar :: StrictTVar m MiniProtocolStatus

data MiniProtocolStatus

Constructors

StatusIdle

StatusStartOnDemand

StatusRunning

data MuxBearer m

Low level access to underlying socket or pipe. There are three smart constructors:

socketAsMuxBearer

pipeAsMuxBearer

Test.Mux.queuesAsMuxBearer

Constructors

MuxBearer

Fields

write :: TimeoutFn m -> MuxSDU -> m Time

Timestamp and send MuxSDU.

read :: TimeoutFn m -> m (MuxSDU, Time)

Read a MuxSDU

sduSize :: SDUSize

Return a suitable MuxSDU payload size.

muxBearerAsChannel :: forall m. Functor m => MuxBearer m -> MiniProtocolNum -> MiniProtocolDir -> Channel m

A channel which wraps each message as an MuxSDU using giving MiniProtocolNum and MiniProtocolDir.

data MuxSDU

Constructors

MuxSDU

Fields

msHeader :: !MuxSDUHeader

msBlob :: !ByteString

data MuxSDUHeader

Constructors

MuxSDUHeader

Fields

mhTimestamp :: !RemoteClockModel

mhNum :: !MiniProtocolNum

mhDir :: !MiniProtocolDir

mhLength :: !Word16

newtype SDUSize

Constructors

SDUSize

Fields

getSDUSize :: Word16

msTimestamp :: MuxSDU -> RemoteClockModel

setTimestamp :: MuxSDU -> RemoteClockModel -> MuxSDU

msNum :: MuxSDU -> MiniProtocolNum

msDir :: MuxSDU -> MiniProtocolDir

msLength :: MuxSDU -> Word16

newtype RemoteClockModel

Constructors

RemoteClockModel

Fields

unRemoteClockModel :: Word32

remoteClockPrecision :: DiffTime

The DiffTime represented by a tick in the RemoteClockModel

data MuxRuntimeError

Constructors

ProtocolAlreadyRunning !MiniProtocolNum !MiniProtocolDir !MiniProtocolStatus

UnknownProtocolInternalError !MiniProtocolNum !MiniProtocolDir

MuxBlockedOnCompletionVar !MiniProtocolNum