Network.Mux.Types

Documentation

data MiniProtocolLimits

Per Miniprotocol limits

Constructors

MiniProtocolLimits

Fields

maximumIngressQueue ∷ !Int64

Limit on the maximum number of bytes that can be queued in the miniprotocol's ingress queue.

newtype MiniProtocolNum

The wire format includes the protocol numbers, and it's vital that these are stable. They are not necessarily dense however, as new ones are added and some old ones retired. So we use a dedicated class for this rather than reusing Enum. This also covers unrecognised protocol numbers on the decoding side.

Constructors

MiniProtocolNum Word16

data MiniProtocolMode

Constructors

ModeInitiator

ModeResponder

data AppType where

Constructors

InitiatorApp ∷ AppType

ResponderApp ∷ AppType

InitiatorAndResponderApp ∷ AppType

type family HasInitiator (appType ∷ AppType) ∷ Bool where ...

Equations

HasInitiator InitiatorApp = True

HasInitiator ResponderApp = False

HasInitiator InitiatorAndResponderApp = True

type family HasResponder (appType ∷ AppType) ∷ Bool where ...

Equations

HasResponder InitiatorApp = False

HasResponder ResponderApp = True

HasResponder InitiatorAndResponderApp = True

newtype MuxApplication (appType ∷ AppType) m a b

Application run by mux layer.

enumeration of client application, e.g. a wallet application communicating with a node using ChainSync and TxSubmission protocols; this only requires to run client side of each protocol.

enumeration of server applications: this application type is mostly useful tests.

enumeration of both client and server applications, e.g. a full node serving downstream peers using server side of each protocol and getting updates from upstream peers using client side of each of the protocols.

Constructors

MuxApplication [MuxMiniProtocol appType m a b]

data MuxMiniProtocol (appType ∷ AppType) m a b

Constructors

MuxMiniProtocol

Fields

miniProtocolNum ∷ !MiniProtocolNum

miniProtocolLimits ∷ !MiniProtocolLimits

miniProtocolRun ∷ !(RunMiniProtocol appType m a b)

data RunMiniProtocol (appType ∷ AppType) m a b where

Constructors

InitiatorProtocolOnly ∷ (Channel m → m a) → RunMiniProtocol InitiatorApp m a Void

ResponderProtocolOnly ∷ (Channel m → m b) → RunMiniProtocol ResponderApp m Void b

InitiatorAndResponderProtocol ∷ (Channel m → m a) → (Channel m → m b) → RunMiniProtocol InitiatorAndResponderApp m a b

data MiniProtocolIx

The index of a protocol in a MuxApplication, used for array indicies

data MuxBearer m

Low level access to underlying socket or pipe. There are three smart constructors:

socketAsMuxBearer

pipeAsMuxBearer

Test.Mux.queuesAsMuxBearer

Constructors

MuxBearer

Fields

write ∷ MuxSDU → m Time

Timestamp and send MuxSDU.

read ∷ m (MuxSDU, Time)

Read a MuxSDU

sduSize ∷ Word16

Return a suitable MuxSDU payload size.

muxBearerAsChannel ∷ MuxBearer IO → MiniProtocolNum → MiniProtocolMode → Channel IO

A channel which wraps each message as an MuxSDU using giving MiniProtocolNum and MiniProtocolMode.

data MuxSDU

Constructors

MuxSDU

Fields

msTimestamp ∷ !RemoteClockModel

msNum ∷ !MiniProtocolNum

msMode ∷ !MiniProtocolMode

msLength ∷ !Word16

msBlob ∷ !ByteString

newtype RemoteClockModel

Constructors

RemoteClockModel

Fields

unRemoteClockModel ∷ Word32

remoteClockPrecision ∷ DiffTime

The DiffTime represented by a tick in the RemoteClockModel